+++ src/main/java/Game/Board/TargetBoardState.java
@@ -6,7 +6,7 @@ public class TargetBoardState extends DefaultBoardState {
 
     public TargetBoardState(int size) {
         super(size);
-        int id = 0;
+        short id = 0;
         for (int y = 0; y < size; y++) {
             for (int x = 0; x < size; x++) {
                 if (y == size - 1 && x == size - 1) {
